+++ social_core/backends/slack.py
@@ -3,8 +3,6 @@ Slack OAuth2 backend, docs at:
     http://psa.matiasaguirre.net/docs/backends/slack.html
     https://api.slack.com/docs/oauth
 """
-import re
-
 from .oauth import BaseOAuth2
 
 
@@ -14,6 +12,7 @@ class SlackOAuth2(BaseOAuth2):
     AUTHORIZATION_URL = 'https://slack.com/oauth/authorize'
     ACCESS_TOKEN_URL = 'https://slack.com/api/oauth.access'
     ACCESS_TOKEN_METHOD = 'POST'
+    DEFAULT_SCOPE = ['identity.basic', 'identity.email']
     SCOPE_SEPARATOR = ','
     REDIRECT_STATE = False
     EXTRA_DATA = [
@@ -26,41 +25,25 @@ class SlackOAuth2(BaseOAuth2):
         """Return user details from Slack account"""
         # Build the username with the team $username@$team_url
         # Necessary to get unique names for all of slack
-        username = response.get('user')
-        if self.setting('USERNAME_WITH_TEAM', True):
-            match = re.search(r'//([^.]+)\.slack\.com', response['url'])
-            username = '{0}@{1}'.format(username, match.group(1))
+        user = response['user']
+        team = response.get('team')
+        name = user['name']
+        email = user.get('email')
+        username = email and email.split('@', 1)[0] or name
+        fullname, first_name, last_name = self.get_user_names(name)
 
-        out = {'username': username}
-        if 'profile' in response:
-            out.update({
-                'email': response['profile'].get('email'),
-                'fullname': response['profile'].get('real_name'),
-                'first_name': response['profile'].get('first_name'),
-                'last_name': response['profile'].get('last_name')
-            })
-        return out
+        if self.setting('USERNAME_WITH_TEAM', True) and team and 'name' in team:
+            name = '{0}@{1}'.format(name, response['team']['name'])
+
+        return {
+            'username': username,
+            'email': email,
+            'fullname': fullname,
+            'first_name': first_name,
+            'last_name': last_name
+        }
 
     def user_data(self, access_token, *args, **kwargs):
         """Loads user data from service"""
-        # Has to be two calls, because the users.info requires a username,
-        # And we want the team information. Check auth.test details at:
-        #   https://api.slack.com/methods/auth.test
-        auth_test = self.get_json('https://slack.com/api/auth.test', params={
-            'token': access_token
-        })
-
-        # https://api.slack.com/methods/users.info
-        user_info = self.get_json('https://slack.com/api/users.info', params={
-            'token': access_token,
-            'user': auth_test.get('user_id')
-        })
-        if user_info.get('user'):
-            # Capture the user data, if available based on the scope
-            auth_test.update(user_info['user'])
-
-        # Clean up user_id vs id
-        auth_test['id'] = auth_test['user_id']
-        auth_test.pop('ok', None)
-        auth_test.pop('user_id', None)
-        return auth_test
+        return self.get_json('https://slack.com/api/users.identity',
+                             params={'token': access_token})